# snakeberry/openapi/mopc_events.yml
openapi: "3.0.2"
info:
  title: MOPC Events API
  description: This API provides interface for CRUD operations with various events
  version: "0.1"
servers:
  - url: https://mopc.xyz
  - url: https://stage.mopc.xyz
  - url: localhost:5001
paths:
  /events:
    get:
      description: Return events list for given dates span
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "400":
          description: Wrong timespan passed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 400
                message: The start date must be less than or equal to the end date
        "5XX":
          $ref: "#/components/responses/UnexpectedError"
      parameters:
        - name: startDate
          in: query
          description: the date FROM which the events will be returned (inclusive)
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: the date by which the events will be returned (inclusive)
          required: true
          schema:
            type: string
            format: date

    post:
      description: Creates a new event. Duplicate names are not allowed
      requestBody:
        description: Event to add
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewEvent"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Incorrect value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 400
                message: The start date must be less than or equal to the end date
        "409":
          description: Name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: 409
                message: Event with name EventName already exists
        "5XX":
          $ref: "#/components/responses/UnexpectedError"

  /events/{id}:
    get:
      description: Returns an event by its ID
      parameters:
        - name: id
          in: path
          description: ID of event to fetch
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Fetched event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          $ref: "#/components/responses/NotFound"
        "5XX":
          $ref: "#/components/responses/UnexpectedError"

    delete:
      description: Deletes a single event based on the ID supplied
      parameters:
        - name: id
          in: path
          description: ID of eventpet to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Event deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "5XX":
          $ref: "#/components/responses/UnexpectedError"

components:
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 404
            message: Not found
    UnexpectedError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: 500
            message: Internal server error

  schemas:
    Event:
      type: object
      required:
        - id
        - name
        - type
        - description
        - startDate
        - endDate
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date

    NewEvent:
      type: object
      required:
        - name
        - type
        - description
        - startDate
        - endDate
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string