# snakeberry/src/events_server/models/event.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class Event(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    Event - a model defined in OpenAPI

        id: The id of this Event.
        name: The name of this Event.
        type: The type of this Event.
        description: The description of this Event.
        start_date: The start_date of this Event.
        end_date: The end_date of this Event.
    """

    id: int = Field(alias="id")
    name: str = Field(alias="name")
    type: str = Field(alias="type")
    description: str = Field(alias="description")
    start_date: date = Field(alias="startDate")
    end_date: date = Field(alias="endDate")

Event.update_forward_refs()