# snakeberry/src/events_server/models/new_event.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class NewEvent(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    NewEvent - a model defined in OpenAPI

        name: The name of this NewEvent.
        type: The type of this NewEvent.
        description: The description of this NewEvent.
        start_date: The start_date of this NewEvent.
        end_date: The end_date of this NewEvent.
    """

    name: str = Field(alias="name")
    type: str = Field(alias="type")
    description: str = Field(alias="description")
    start_date: date = Field(alias="startDate")
    end_date: date = Field(alias="endDate")

NewEvent.update_forward_refs()